src/test/java/com/bitc/shop/service/CartServiceTest.java
package com.bitc.shop.service;

import com.bitc.shop.constant.ItemSellStatus;
import com.bitc.shop.dto.CartItemDto;
import com.bitc.shop.entity.CartItem;
import com.bitc.shop.entity.Item;
import com.bitc.shop.entity.Member;
import com.bitc.shop.repository.CartItemRepository;
import com.bitc.shop.repository.ItemRepository;
import com.bitc.shop.repository.MemberRepository;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.TestPropertySource;
import org.springframework.transaction.annotation.Transactional;

import javax.persistence.EntityNotFoundException;

import static org.junit.jupiter.api.Assertions.assertEquals;

@SpringBootTest
@Transactional
@TestPropertySource(locations = "classpath:application-test.properties")
public class CartServiceTest {

  @Autowired
  ItemRepository itemRepository;

  @Autowired
  MemberRepository memberRepository;

  @Autowired
  CartService cartService;

  @Autowired
  CartItemRepository cartItemRepository;

  public Item saveItem() {
    Item item = new Item();
    item.setItemNm("테스트 상품");
    item.setPrice(10000);
    item.setItemDetail("테스트 상품 상세 설명");
    item.setItemSellStatus(ItemSellStatus.SELL);
    item.setStockNumber(100);

    return itemRepository.save(item);
  }

  public Member saveMember() {
    Member member = new Member();
    member.setEmail("test@test.com");

    return memberRepository.save(member);
  }

  @Test
  @DisplayName("장바구니 담기 테스트")
  public void addCart() {
    Item item = saveItem();
    Member member = saveMember();

    CartItemDto cartItemDto = new CartItemDto();
    cartItemDto.setCount(5);
    cartItemDto.setItemId(item.getId());

    Long cartItemId = cartService.addCart(cartItemDto, member.getEmail());

    CartItem cartItem = cartItemRepository.findById(cartItemId)
        .orElseThrow(EntityNotFoundException::new);

    assertEquals(item.getId(), cartItem.getItem().getId());
    assertEquals(cartItemDto.getCount(), cartItem.getCount());
  }
}
